0.9.0

+++ routes/api/permission-group.php
@@ -1,70 +1,18 @@
 <?php
 
-use Illuminate\Support\Facades\Route;
+use App\Http\Controllers\RBAC\PermissionGroupController;
 
 Route::group(
     [
         'as' => 'permission:group:',
         'namespace' => 'App\Http\Controllers\RBAC',
-        'middleware' => [],
     ],
     function () {
-        Route::get(
-            '/permission-group',
-            [
-                'as' => 'all',
-                'uses' => 'PermissionGroupController@index',
-            ]
-        );
-
-        Route::get(
-            '/permission-group/{id}',
-            [
-                'as' => 'show',
-                'uses' => 'PermissionGroupController@show',
-            ]
-        );
-
-        Route::get(
-            '/permission-group/search',
-            [
-                'as' => 'search',
-                'uses' => 'PermissionGroupController@search',
-            ]
-        );
-    }
-);
-
-Route::group(
-    [
-        'as' => 'permission:group:',
-        'namespace' => 'App\Http\Controllers\RBAC',
-        'middleware' => ['auth:api'],
-    ],
-    function () {
-
-        Route::delete(
-            '/permission-group/{id}',
-            [
-                'as' => 'delete',
-                'uses' => 'PermissionGroupController@destroy',
-            ]
-        );
-
-        Route::post(
-            '/permission-group',
-            [
-                'as' => 'create',
-                'uses' => 'PermissionGroupController@store',
-            ]
-        );
-
-        Route::put(
-            '/permission-group/{id}',
-            [
-                'as' => 'update',
-                'uses' => 'PermissionGroupController@update',
-            ]
-        );
+        Route::get('/permission-group',         [PermissionGroupController::class, 'index'])->name('all');
+        Route::get('/permission-group/{id}',    [PermissionGroupController::class, 'show'])->name('show');
+        Route::get('/permission-group/search',  [PermissionGroupController::class, 'search'])->name('search');
+        Route::delete('/permission-group/{id}', [PermissionGroupController::class, 'destroy'])->name('delete');
+        Route::post('/permission-group',        [PermissionGroupController::class, 'store'])->name('create');
+        Route::put('/permission-group/{id}',    [PermissionGroupController::class, 'update'])->name('update');
     }
 );
