+++ routes/api/role.php
@@ -1,39 +1,6 @@
 <?php
 
-use Illuminate\Support\Facades\Route;
-
-Route::group(
-    [
-        'as' => 'role:',
-        'namespace' => 'App\Http\Controllers\RBAC',
-        'middleware' => [],
-    ],
-    function () {
-        Route::get(
-            '/role',
-            [
-                'as' => 'all',
-                'uses' => 'RoleController@index',
-            ]
-        );
-
-        Route::get(
-            '/role/{id}',
-            [
-                'as' => 'show',
-                'uses' => 'RoleController@show',
-            ]
-        );
-
-        Route::get(
-            '/role/search',
-            [
-                'as' => 'search',
-                'uses' => 'RoleController@search',
-            ]
-        );
-    }
-);
+use App\Http\Controllers\RBAC\RoleController;
 
 Route::group(
     [
@@ -42,45 +9,13 @@ Route::group(
         'middleware' => ['auth:api'],
     ],
     function () {
-
-        Route::post(
-            '/role/{roleId}/{permissionId}',
-            [
-                'as' => 'attachPermission',
-                'uses' => 'RoleController@attachPermission',
-            ]
-        );
-
-        Route::delete(
-            '/role/{roleId}/{permissionId}',
-            [
-                'as' => 'detachPermission',
-                'uses' => 'RoleController@detachPermission',
-            ]
-        );
-
-        Route::delete(
-            '/role/{id}',
-            [
-                'as' => 'delete',
-                'uses' => 'RoleController@destroy',
-            ]
-        );
-
-        Route::post(
-            '/role',
-            [
-                'as' => 'create',
-                'uses' => 'RoleController@store',
-            ]
-        );
-
-        Route::put(
-            '/role/{id}',
-            [
-                'as' => 'update',
-                'uses' => 'RoleController@update',
-            ]
-        );
+        Route::get('/role',         [RoleController::class, 'index'])->name('all');
+        Route::get('/role/{id}',    [RoleController::class, 'show'])->name('show');
+        Route::get('/role/search',  [RoleController::class, 'search'])->name('search');
+        Route::post('/role/{roleId}/{permissionId}', [RoleController::class, 'attachPermission'])->name('attachPermission');
+        Route::delete('/role/{roleId}/{permissionId}', [RoleController::class, 'detachPermission'])->name('detachPermission');
+        Route::delete('/role/{id}', [RoleController::class, 'destroy'])->name('delete');
+        Route::post('/role',        [RoleController::class, 'store'])->name('create');
+        Route::put('/role/{id}',    [RoleController::class, 'update'])->name('update');
     }
 );
